Moved the apollo extensions to its own repo (swift package) and extended the model to include subscriptions. This means the server implementation now stores messages which can be created in the app. Multiple clients will be able to retrieve these changes and have it visible as soon as these items are added, the problem is apollo only caches items as you fetch them and by default, the configuration is to not fetch if there is anything in the cache.
The DataStore changed a bit, I started including a bunch of functions there, but I am not using them (at least not the flush and listen to messages), but committing those at the moment as I have increased this change to the point I do not like it.

Network was rewritten because I don't like those lazy properties everywhere, plus in order to include subscriptions the endpoint needs to be recreated with the appropriate auth token.

Note:
Playing with code for days and not committing is bad.